Moving http client functions
