+++ kong-0.0.1beta-1.rockspec
@@ -44,6 +44,7 @@ build = {
     ["kong.tools.utils"] = "src/kong/tools/utils.lua",
     ["kong.tools.timestamp"] = "src/kong/tools/timestamp.lua",
     ["kong.tools.cache"] = "src/kong/tools/cache.lua",
+    ["kong.tools.http_client"] = "src/kong/tools/http_client.lua",
     ["kong.tools.faker"] = "src/kong/tools/faker.lua",
     ["kong.tools.migrations"] = "src/kong/tools/migrations.lua",
