+++ spec/proxy/authentication_plugin_spec.lua
@@ -1,5 +1,5 @@
 local spec_helper = require "spec.spec_helpers"
-local utils = require "kong.tools.utils"
+local http_client = require "kong.tools.http_client"
 local cjson = require "cjson"
 
 local kProxyURL = spec_helper.PROXY_URL
@@ -21,35 +21,35 @@ describe("Authentication Plugin #proxy", function()
   describe("Query Authentication", function()
 
     it("should return invalid credentials when the credential value is wrong", function()
-      local response, status, headers = utils.get(kGetURL, {apikey = "asd"}, {host = "test.com"})
+      local response, status, headers = http_client.get(kGetURL, {apikey = "asd"}, {host = "test.com"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in GET", function()
-      local response, status, headers = utils.get(kGetURL, {apikey123 = "apikey123"}, {host = "test.com"})
+      local response, status, headers = http_client.get(kGetURL, {apikey123 = "apikey123"}, {host = "test.com"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in POST", function()
-      local response, status, headers = utils.post(kPostURL, {apikey123 = "apikey123"}, {host = "test.com"})
+      local response, status, headers = http_client.post(kPostURL, {apikey123 = "apikey123"}, {host = "test.com"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should pass with GET", function()
-      local response, status, headers = utils.get(kGetURL, {apikey = "apikey123"}, {host = "test.com"})
+      local response, status, headers = http_client.get(kGetURL, {apikey = "apikey123"}, {host = "test.com"})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("apikey123", parsed_response.args.apikey)
     end)
 
     it("should pass with POST", function()
-      local response, status, headers = utils.post(kPostURL, {apikey = "apikey123"}, {host = "test.com"})
+      local response, status, headers = http_client.post(kPostURL, {apikey = "apikey123"}, {host = "test.com"})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("apikey123", parsed_response.form.apikey)
@@ -60,35 +60,35 @@ describe("Authentication Plugin #proxy", function()
   describe("Header Authentication", function()
 
     it("should return invalid credentials when the credential value is wrong", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test2.com", apikey = "asd"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test2.com", apikey = "asd"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in GET", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test2.com", apikey123 = "apikey123"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test2.com", apikey123 = "apikey123"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in POST", function()
-      local response, status, headers = utils.post(kPostURL, {}, {host = "test2.com", apikey123 = "apikey123"})
+      local response, status, headers = http_client.post(kPostURL, {}, {host = "test2.com", apikey123 = "apikey123"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should pass with GET", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test2.com", apikey = "apikey123"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test2.com", apikey = "apikey123"})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("apikey123", parsed_response.headers.Apikey)
     end)
 
     it("should pass with POST", function()
-      local response, status, headers = utils.post(kPostURL, {}, {host = "test2.com", apikey = "apikey123"})
+      local response, status, headers = http_client.post(kPostURL, {}, {host = "test2.com", apikey = "apikey123"})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("apikey123", parsed_response.headers.Apikey)
@@ -99,49 +99,49 @@ describe("Authentication Plugin #proxy", function()
   describe("Basic Authentication", function()
 
     it("should return invalid credentials when the credential value is wrong", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test3.com", authorization = "asd"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test3.com", authorization = "asd"})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should not pass when passing only the password", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test3.com", authorization = "Basic OmFwaWtleTEyMw=="})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test3.com", authorization = "Basic OmFwaWtleTEyMw=="})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should not pass when passing only the username", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test3.com", authorization = "Basic dXNlcjEyMzo="})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test3.com", authorization = "Basic dXNlcjEyMzo="})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in GET", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test3.com", authorization123 = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test3.com", authorization123 = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should return invalid credentials when the credential parameter name is wrong in POST", function()
-      local response, status, headers = utils.post(kPostURL, {}, {host = "test3.com", authorization123 = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
+      local response, status, headers = http_client.post(kPostURL, {}, {host = "test3.com", authorization123 = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
       local body = cjson.decode(response)
       assert.are.equal(403, status)
       assert.are.equal("Your authentication credentials are invalid", body.message)
     end)
 
     it("should pass with GET", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test3.com", authorization = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test3.com", authorization = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("Basic dXNlcm5hbWU6cGFzc3dvcmQ=", parsed_response.headers.Authorization)
     end)
 
     it("should pass with POST", function()
-      local response, status, headers = utils.post(kPostURL, {}, {host = "test3.com", authorization = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
+      local response, status, headers = http_client.post(kPostURL, {}, {host = "test3.com", authorization = "Basic dXNlcm5hbWU6cGFzc3dvcmQ="})
       assert.are.equal(200, status)
       local parsed_response = cjson.decode(response)
       assert.are.equal("Basic dXNlcm5hbWU6cGFzc3dvcmQ=", parsed_response.headers.Authorization)
