+++ spec/proxy/core_access_spec.lua
@@ -1,5 +1,5 @@
 local spec_helper = require "spec.spec_helpers"
-local utils = require "kong.tools.utils"
+local http_client = require "kong.tools.http_client"
 local cjson = require "cjson"
 
 local kProxyURL = spec_helper.PROXY_URL
@@ -21,7 +21,7 @@ describe("Proxy API #proxy", function()
   describe("Invalid API", function()
 
     it("should return API not found when the API is missing", function()
-      local response, status, headers = utils.get(kProxyURL)
+      local response, status, headers = http_client.get(kProxyURL)
       local body = cjson.decode(response)
       assert.are.equal(404, status)
       assert.are.equal("API not found", body.message)
@@ -32,7 +32,7 @@ describe("Proxy API #proxy", function()
   describe("Existing API", function()
 
     it("should return API found when the API has been created", function()
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test4.com"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test4.com"})
       assert.are.equal(200, status)
     end)
