+++ spec/proxy/ratelimiting_plugin_spec.lua
@@ -1,5 +1,5 @@
 local spec_helper = require "spec.spec_helpers"
-local utils = require "kong.tools.utils"
+local http_client = require "kong.tools.http_client"
 local cjson = require "cjson"
 
 local kProxyURL = spec_helper.PROXY_URL
@@ -24,14 +24,14 @@ describe("RateLimiting Plugin #proxy", function()
       local limit = 2
 
       for i = 1, limit do
-        local response, status, headers = utils.get(kGetURL, {}, {host = "test5.com"})
+        local response, status, headers = http_client.get(kGetURL, {}, {host = "test5.com"})
         assert.are.equal(200, status)
         assert.are.same(tostring(limit), headers["x-ratelimit-limit"])
         assert.are.same(tostring(limit - i), headers["x-ratelimit-remaining"])
       end
 
       -- Third query, while limit is 2/minute
-      local response, status, headers = utils.get(kGetURL, {}, {host = "test5.com"})
+      local response, status, headers = http_client.get(kGetURL, {}, {host = "test5.com"})
       local body = cjson.decode(response)
       assert.are.equal(429, status)
       assert.are.equal("API rate limit exceeded", body.message)
@@ -48,14 +48,14 @@ describe("RateLimiting Plugin #proxy", function()
         local limit = 2
 
         for i = 1, limit do
-          local response, status, headers = utils.get(kGetURL, {apikey = "apikey122"}, {host = "test6.com"})
+          local response, status, headers = http_client.get(kGetURL, {apikey = "apikey122"}, {host = "test6.com"})
           assert.are.equal(200, status)
           assert.are.same(tostring(limit), headers["x-ratelimit-limit"])
           assert.are.same(tostring(limit - i), headers["x-ratelimit-remaining"])
         end
 
         -- Third query, while limit is 2/minute
-        local response, status, headers = utils.get(kGetURL, {apikey = "apikey122"}, {host = "test6.com"})
+        local response, status, headers = http_client.get(kGetURL, {apikey = "apikey122"}, {host = "test6.com"})
         local body = cjson.decode(response)
         assert.are.equal(429, status)
         assert.are.equal("API rate limit exceeded", body.message)
@@ -70,13 +70,13 @@ describe("RateLimiting Plugin #proxy", function()
         local limit = 4
 
         for i = 1, limit do
-          local response, status, headers = utils.get(kGetURL, {apikey = "apikey123"}, {host = "test6.com"})
+          local response, status, headers = http_client.get(kGetURL, {apikey = "apikey123"}, {host = "test6.com"})
           assert.are.equal(200, status)
           assert.are.same(tostring(limit), headers["x-ratelimit-limit"])
           assert.are.same(tostring(limit - i), headers["x-ratelimit-remaining"])
         end
 
-        local response, status, headers = utils.get(kGetURL, {apikey = "apikey123"}, {host = "test6.com"})
+        local response, status, headers = http_client.get(kGetURL, {apikey = "apikey123"}, {host = "test6.com"})
         local body = cjson.decode(response)
         assert.are.equal(429, status)
         assert.are.equal("API rate limit exceeded", body.message)
