+++ src/kong/tools/utils.lua
@@ -1,9 +1,6 @@
 local constants = require "kong.constants"
 local cjson = require "cjson"
-local ltn12 = require "ltn12"
 local yaml = require "yaml"
-local http = require "socket.http"
-local url = require "socket.url"
 local lfs = require "lfs"
 
 local _M = {}
@@ -183,69 +180,6 @@ function _M.create_timer(func, data)
 end
 
 --
--- Cache utils
---
-function _M.cache_set(key, value, exptime)
-  if exptime == nil then exptime = 0 end -- By default never expire
-  local cache = ngx.shared.cache
-  if value then
-    value = cjson.encode(value)
-  end
-  if ngx then
-    ngx.log(ngx.DEBUG, " saving cache key \""..key.."\": "..value)
-  end
-  local succ, err, forcible = cache:set(key, value, exptime)
-  return succ, err, forcible
-end
-
-function _M.cache_get(key)
-  if ngx then
-    ngx.log(ngx.DEBUG, " Try to get cache key \""..key.."\"")
-  end
-
-  local cache = ngx.shared.cache
-  local value, flags = cache:get(key)
-  if value then
-    if ngx then
-      ngx.log(ngx.DEBUG, " Found cache value for key \""..key.."\": "..value)
-    end
-    value = cjson.decode(value)
-  end
-  return value, flags
-end
-
-function _M.cache_delete(key)
-  local cache = ngx.shared.cache
-  cache:delete(key)
-end
-
-function _M.cache_api_key(host)
-  return constants.CACHE.APIS.."/"..host
-end
-
-function _M.cache_plugin_key(name, api_id, application_id)
-  return constants.CACHE.PLUGINS.."/"..name.."/"..api_id..(application_id and "/"..application_id or "")
-end
-
-function _M.cache_application_key(public_key)
-  return constants.CACHE.APPLICATIONS.."/"..public_key
-end
-
-function _M.cache_get_and_set(key, cb)
-  local val = _M.cache_get(key)
-  if not val then
-    val = cb()
-    if val then
-      local succ, err = _M.cache_set(key, val)
-      if not succ and ngx then
-        ngx.log(ngx.ERR, err)
-      end
-    end
-  end
-  return val
-end
-
---
 -- Disk I/O utils
 --
 function _M.read_file(path)
@@ -279,131 +213,6 @@ function _M.retrieve_files(path, pattern)
   end
 
   return files
-end
-
---
--- HTTP calls utils
---
-
--- Builds a querystring from a table, separated by `&`
--- @param tab The key/value parameters
--- @param key The parent key if the value is multi-dimensional (optional)
--- @return a string representing the built querystring
-function _M.build_query(tab, key)
-  if ngx then
-    return ngx.encode_args(tab)
-  else
-    local query = {}
-    local keys = {}
-
-    for k in pairs(tab) do
-      keys[#keys+1] = k
-    end
-
-    table.sort(keys)
-
-    for _,name in ipairs(keys) do
-      local value = tab[name]
-      if key then
-        name = string.format("%s[%s]", tostring(key), tostring(name))
-      end
-      if type(value) == "table" then
-        query[#query+1] = _M.build_query(value, name)
-      else
-        value = tostring(value)
-        if value ~= "" then
-          query[#query+1] = string.format("%s=%s", name, value)
-        else
-          query[#query+1] = name
-        end
-      end
-    end
-
-    return table.concat(query, "&")
-  end
-end
-
-local function http_call(options)
-  -- Set Host header accordingly
-  if not options.headers["host"] then
-    local parsed_url = url.parse(options.url)
-    local port_segment = ""
-    if parsed_url.port then
-      port_segment = ":" .. parsed_url.port
-    end
-    options.headers["host"] = parsed_url.host .. port_segment
-  end
-
-  -- Returns: response, code, headers
-  local resp = {}
-  options.sink = ltn12.sink.table(resp)
-
-  local _, code, headers = http.request(options)
-  return resp[1], code, headers
-end
-
--- GET methpd
-function _M.get(url, querystring, headers)
-  if not headers then headers = {} end
-
-  if querystring then
-    url = string.format("%s?%s", url, _M.build_query(querystring))
-  end
-
-  return http_call {
-    method = "GET",
-    url = url,
-    headers = headers
-  }
-end
-
--- POST methpd
-function _M.post(url, form, headers)
-  if not headers then headers = {} end
-  if not form then form = {} end
-
-  local body = _M.build_query(form)
-  headers["content-length"] = string.len(body)
-  headers["content-type"] = "application/x-www-form-urlencoded"
-
-  return http_call {
-    method = "POST",
-    url = url,
-    headers = headers,
-    source = ltn12.source.string(body)
-  }
-end
-
--- PUT method
-function _M.put(url, table, headers)
-  if not headers then headers = {} end
-  if not table then table = {} end
-  local raw_json = cjson.encode(table)
-
-  headers["content-length"] = string.len(raw_json)
-  headers["content-type"] = "application/json"
-
-  return http_call {
-    method = "PUT",
-    url = url,
-    headers = headers,
-    source = ltn12.source.string(raw_json)
-  }
-end
-
--- DELETE methpd
-function _M.delete(url, querystring, headers)
-  if not headers then headers = {} end
-
-  if querystring then
-    url = string.format("%s?%s", url, _M.build_query(querystring))
-  end
-
-  return http_call {
-    method = "DELETE",
-    url = url,
-    headers = headers
-  }
 end
 
 --
